Storybook: Cleanup and simplify Tooltip stories (#3952)
